convert_to_dsfr

permet de convertir une application shiny en une application shiny suivant le DSFR

convert_to_dsfr(path = "R/", version = get_dsfr_version())

Arguments

path

le chemin du dossier sur lequel faire la conversion, par defaut "R/"

version

le numero de version du dsfr a utiliser, par exemple "1.7.2". Par défaut, la derniere disponible. Utilisé plutot lors du developpement pour tester les retrocompatibilites.

Value

des fichiers modifies

Examples

mydir <- tempfile(pattern = "app")
dir.create(mydir)
golem::create_golem(
  mydir,
  overwrite = TRUE,
  open = FALSE
)
#> ── Checking package name ───────────────────────────────────────────────────────
#> ✔ Valid package name
#> • Overwriting existing project.
#> ── Copying package skeleton ────────────────────────────────────────────────────
#> ✔ Copied app skeleton
#> ── Setting the default config ──────────────────────────────────────────────────
#> ✔ Configured app
#> ── Running project hook function ───────────────────────────────────────────────
#> ✔ All set
#> ✔ Setting active project to '/tmp/RtmpIvKh6K/app3ffc7a53d3f3'
#> ── Done ────────────────────────────────────────────────────────────────────────
#> A new golem named app3ffc7a53d3f3 was created at /tmp/RtmpIvKh6K/app3ffc7a53d3f3 .
#> To continue working on your app, start editing the 01_start.R file.
convert_to_dsfr(path = file.path(mydir, "R"))
#> scan /tmp/RtmpIvKh6K/app3ffc7a53d3f3/R/app_config.R
#> scan /tmp/RtmpIvKh6K/app3ffc7a53d3f3/R/app_server.R
#> scan /tmp/RtmpIvKh6K/app3ffc7a53d3f3/R/app_ui.R
#> scan /tmp/RtmpIvKh6K/app3ffc7a53d3f3/R/run_app.R
#> la dependance a shinygouv a ete ajoute dans app_ui.R
#> [1] "Conversion terminee"