footer_bottom_dsfr

Il se compose des éléments suivants:

la liste de liens liés aux obligations légales. Cette liste doit être définie en fonction du site, toutefois les liens & contenus suivants sont obligatoires : “accessibilité : non/partiellement/totalement conforme”, mentions légales, données personnelles et gestion des cookies. - obligatoire.

une mention de la licence - obligatoire “Sauf mention contraire, tous les contenus de ce site sont sous licence etalab-2.0”

Il est possible d'ajouter des url supplémentaires dans le corps et le bas du pied de page.

Permet d'ajouter des liends/url dans le corps du pied de page

Permet d'ajouter des liends/url dans le bas du pied de page

Permet de créer la balise html pour ajouter un lien en fonction de sa destination (body/bottom)

Permet d'ajouter des liends/url dans le bas du pied de page

footer_bottom_dsfr(
  accessibilite = "non",
  ...,
  lien_accessibilite = "#",
  lien_mentions_legales = "#",
  lien_donnees_personnelles = "#",
  lien_gestion_cookies = "#"
)

footer_dsfr(
  intitule = NULL,
  description = NULL,
  accessibilite = "non",
  lien_accessibilite = "#",
  lien_mentions_legales = "#",
  lien_donnees_personnelles = "#",
  lien_gestion_cookies = "#",
  add_extra_url_body = NULL,
  add_extra_url_bottom = NULL
)

footer_extra_url_body_dsfr(...)

footer_extra_url_bottom_dsfr(...)

footer_li_dsfr(id = NULL, href, titre = NULL, texte, type = "body")

footer_extra_url_bottom_dsfr(...)

Arguments

accessibilite

le type d'accessibilité (non/partiellement/totalement conforme)

...

d'autre liens de références (voir footer_li_dsfr())

lien_accessibilite

lien url vers l'accessibilite

lien_mentions_legales

lien url vers les mentions légales

lien_donnees_personnelles

lien url vers les données personnelles

lien_gestion_cookies

lien url vers la gestion des cookies

intitule

la partie "intitule" du bloc marque. Utiliser un shiny::tags$span()pour mettre l'intitulé sur plusieurs lignes ou un vecteur.

description

un texte de présentation - optionnel

add_extra_url_body

url supplementaire a ajouter au corps du pied de page (voir footer_extra_url_body_dsfr())

add_extra_url_bottom

url supplementaire a ajouter au bas du pied de page (voir footer_extra_url_bottom_dsfr())

id

un identifiant (optionnel)

href

une url

titre

le titre de l infobulle (optionnel)

texte

le texte

type

le type de rendu de l url (body/bottom, par defaut body)

Value

html

html

html

html

html

html

Examples

if (interactive()) {
  library(shiny)
  shinyApp(
    ui = fluidPage_dsfr(
      footer = footer_dsfr(
        intitule = c("Territoire", "Service"),
        description = "une description",
        accessibilite = "non",
        add_extra_url_body = footer_extra_url_body_dsfr(
          footer_li_dsfr(
            href = "https://siteinternet.fr",
            titre = "siteinternet.fr.fr - nouvelle fenêtre",
            texte = "siteinternet.fr.fr", type = "body"
          ),
          footer_li_dsfr(
            href = "https://siteinternet.fr",
            titre = "siteinternet.fr.fr - nouvelle fenêtre",
            texte = "siteinternet.fr.fr", type = "body"
          )
        ),
        add_extra_url_bottom = footer_extra_url_bottom_dsfr(

          footer_li_dsfr(
            href = "https://siteinternet.fr",
            titre = "siteinternet.fr.fr - nouvelle fenêtre",
            texte = "siteinternet.fr.fr", type = "bottom"
          )
        )
      ),
      tabsetPanel_dsfr(
        "coucou",
        tabPanel_dsfr(
          id = "tab1",
          title = "Une prems tab",
          content = tagList(h3("Coucou"),
                            plotOutput("plot"))
        ),
        tabPanel_dsfr(
          id = "tab2",
          title = "Une deuxieme tab",
          content = tagList(h3("Salut"),
                            plotOutput("plot2"))
        )
      )
    ),
    server = function(input, output) {
      output$plot <- renderPlot({
        plot(cars)
      })
      output$plot2 <- renderPlot({
        plot(iris)
      })
    }
  )
}
footer_extra_url_body_dsfr(
                footer_li_dsfr(href = "http://siteinternet.fr", texte = "siteinternet.fr", type = "body"),
                footer_li_dsfr(href = "http://siteinternet.fr", texte = "siteinternet.fr", type = "body"))
#> <li class="fr-footer__content-item">
#>   <a id="footer__content-link__" href="http://siteinternet.fr" target="_blank" title class="fr-footer__content-link">siteinternet.fr</a>
#> </li>
#> <li class="fr-footer__content-item">
#>   <a id="footer__content-link__" href="http://siteinternet.fr" target="_blank" title class="fr-footer__content-link">siteinternet.fr</a>
#> </li>
footer_extra_url_bottom_dsfr(footer_li_dsfr(href = "http://siteinternet.fr", texte = "siteinternet.fr"))
#> <li class="fr-footer__content-item">
#>   <a id="footer__content-link__" href="http://siteinternet.fr" target="_blank" title class="fr-footer__content-link">siteinternet.fr</a>
#> </li>
footer_li_dsfr(href = "http://siteinternet.fr", texte = "siteinternet.fr", type = "body")
#> <li class="fr-footer__content-item">
#>   <a id="footer__content-link__" href="http://siteinternet.fr" target="_blank" title class="fr-footer__content-link">siteinternet.fr</a>
#> </li>
footer_li_dsfr(href = "http://siteinternet.fr", texte = "siteinternet.fr", type = "bottom")
#> <li class="fr-footer__bottom-item">
#>   <a id="footer__bottom-link__" href="http://siteinternet.fr" title class="fr-footer__bottom-link">siteinternet.fr</a>
#> </li>